07 ViewSets & Routers in Django Rest API Bangla
1.What is ViewSets and Routers in Django ?
Tutorial 6: ViewSets & Routers. REST framework includes an abstraction for dealing with ViewSets , that allows the developer to concentrate 
on modeling the state and interactions of the API, and leave the URL construction to be handled automatically, based on common conventions.

2. What is an API?
An application programming interface is a connection between computers or between computer programs. It is a type of software interface, 
offering a service to other pieces of software. A document or standard that describes how to build or use such a connection or interface 
is called an API. A computer system that meets this standard is said to implement or expose an API. The term API may refer either to the 
specification or to the implementation.

3. What is REST API ?
Representational state transfer is a software architectural style that was created to guide the design and development of the architecture 
for the World Wide Web. REST defines a set of constraints for how the architecture of an Internet-scale distributed hypermedia system, such 
as the Web, should behave. The REST architectural style emphasizes the scalability of interactions between components, uniform interfaces, 
independent deployment of components, and the creation of a layered architecture to facilitate caching components to reduce user-perceived 
latency, enforce security, and encapsulate legacy systems.

4. Django REST Framework ?
Django REST framework is a powerful and flexible toolkit for building Web APIs. ... The Web browsable API is a huge usability win for your 
developers. Authentication policies including packages for OAuth1a and OAuth2. Serialization that supports both ORM and non-ORM data sources.

5. how to work Django rest api ?
 1.Set up Django.
 2.Create a model in the database that the Django ORM will manage.
 3.Set up the Django REST Framework.
 4.Serialize the model from step 2.
 5.Create the URI endpoints to view the serialized data.

#ViewSets_&_Routers
#API_Bangla_Tutorial
#REST_API_Bangla_Tutorial